+++ src/ai/models/ai_gateway/__init__.py
@@ -6,12 +6,8 @@ time.  This matters for sandboxed runtimes (e.g. Temporal workflow workers).
 """
 
 from . import errors
-from .types import GenerateParams, ImageParams, VideoParams
 
 __all__ = [
-    "GenerateParams",
-    "ImageParams",
-    "VideoParams",
     "errors",
 ]
